--- src/main/java/MarkDoneCommand.java
@@ -1,20 +1,0 @@
-public class MarkDoneCommand extends Command {
-    private int index;
-
-    public MarkDoneCommand(int index) {
-        this.index = index - 1;
-    }
-
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
-        tasks.markTask(index);
-        ui.printMessage("Beep Boop NICE! I've marked this task as done:\n\t\t" +
-                tasks.get(index));
-        storage.save(tasks.getAllTasks());
-    }
-
-    @Override
-    public boolean isExit() {
-        return false;
-    }
-}
