+++ src/main/java/duke/command/AddDeadlineCommand.java
@@ -1,3 +1,11 @@
+package duke.command;
+
+import duke.exception.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.task.Deadline;
+
 import java.time.LocalDate;
 
 public class AddDeadlineCommand extends Command {
