+++ src/main/java/duke/command/AddToDoCommand.java
@@ -1,3 +1,11 @@
+package duke.command;
+
+import duke.exception.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.task.ToDo;
+
 public class AddToDoCommand extends Command {
     private String description;
