+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.exception.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
 public class DeleteCommand extends Command {
     private int index;
