+++ src/main/java/duke/command/MarkDoneCommand.java
@@ -1,14 +1,21 @@
-public class UnmarkDoneCommand extends Command {
+package duke.command;
+
+import duke.exception.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
+public class MarkDoneCommand extends Command {
     private int index;
 
-    public UnmarkDoneCommand(int index) {
+    public MarkDoneCommand(int index) {
         this.index = index - 1;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
-        tasks.unmarkTask(index);
-        ui.printMessage("Beep Boop NICE! I've unmarked this task as done:\n\t\t" +
+        tasks.markTask(index);
+        ui.printMessage("Beep Boop NICE! I've marked this task as done:\n\t\t" +
                 tasks.get(index));
         storage.save(tasks.getAllTasks());
     }
