+++ src/main/java/duke/command/UnmarkDoneCommand.java
@@ -1,0 +1,27 @@
+package duke.command;
+
+import duke.exception.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
+public class UnmarkDoneCommand extends Command {
+    private int index;
+
+    public UnmarkDoneCommand(int index) {
+        this.index = index - 1;
+    }
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
+        tasks.unmarkTask(index);
+        ui.printMessage("Beep Boop NICE! I've unmarked this task as done:\n\t\t" +
+                tasks.get(index));
+        storage.save(tasks.getAllTasks());
+    }
+
+    @Override
+    public boolean isExit() {
+        return false;
+    }
+}
