+++ src/main/java/duke/task/ToDo.java
@@ -1,5 +1,7 @@
+package duke.task;
+
 public class ToDo extends Task {
-    // Constructor for ToDo
+    // Constructor for duke.task.ToDo
     public ToDo(String description) {
         super(description);
     }
